some changes

+++ src/modules/vehicles/vehicles.service.ts
@@ -1,5 +1,6 @@
 import { Pool } from "pg";
 import config from "../../config";
+import jwt, { JwtPayload } from "jsonwebtoken";
 
 const pool = new Pool({
   connectionString: config.connectionStr,
@@ -13,7 +14,7 @@ type VehiclePayload = {
   availability_status: string;
 };
 
-const postVehicle = async (payload: VehiclePayload) => {
+const postVehicle = async (payload: VehiclePayload , token :string) => {
   const {
     vehicle_name,
     type,
@@ -22,6 +23,11 @@ const postVehicle = async (payload: VehiclePayload) => {
     availability_status,
   } = payload;
 
+   const decoded = jwt.decode(token) as JwtPayload;
+  const role = decoded.role;
+  if (role != "admin") {
+    return "You dont have access to Create Vehicles";
+  }
   try {
     const result = await pool.query(
       `INSERT INTO vehicles(vehicle_name, "type", registration_number, daily_rent_price , availability_status)   VALUES ($1, $2, $3, $4, $5) RETURNING *`,
@@ -97,22 +103,29 @@ const updateVehicle = async (id: Number, payload: Record<string, unknown>) => {
   }
 };
 
-const deleteVehicle = async (id: Number) => {
+const deleteVehicle = async (id: Number, token: string) => {
+  const decoded = jwt.decode(token) as JwtPayload;
+  const role = decoded.role;
+  if (role != "admin") {
+    return "You dont have access to Delete Vehicles";
+  }
   try {
     const test = await pool.query(`SELECT * FROM vehicles WHERE id=$1`, [id]);
 
-    console.log(test)
+    console.log(test);
 
     if (!(test.rows[0].availability_status === "booked")) {
       const result = await pool.query(`DELETE FROM vehicles WHERE id=$1`, [id]);
 
-      console.log(result)
+      console.log(result);
 
       if (result.rowCount === 1) {
         return result.command;
       } else {
         return "Unsuccessfull";
       }
+    }else{
+      return "Vehicle is booked stage!";
     }
   } catch (err) {
     return err;
